feat: Extract releases deploys subcommand to a separate deploys command
