+++ src/commands/files_delete.rs
@@ -9,7 +9,7 @@ use crate::config::Config;
 pub fn make_command(command: Command) -> Command {
     command
         .about("Delete a release file.")
-        // Backwards compatibility with `releases files <VERSION>` commands.
+        // Backward compatibility with `releases files <VERSION>` commands.
         .arg(Arg::new("version").long("version").hide(true))
         .arg(
             Arg::new("names")
