+++ src/commands/files_list.rs
@@ -7,7 +7,7 @@ use crate::{api::Api, config::Config, utils::formatting::Table};
 pub fn make_command(command: Command) -> Command {
     command
         .about("List all release files.")
-        // Backwards compatibility with `releases files <VERSION>` commands.
+        // Backward compatibility with `releases files <VERSION>` commands.
         .arg(Arg::new("version").long("version").hide(true))
 }
