+++ src/config.rs
@@ -294,7 +294,7 @@ impl Config {
             .ok_or_else(|| format_err!("A release slug is required (provide with --release)"))
     }
 
-    // Backwards compatibility with `releases files <VERSION>` commands.
+    // Backward compatibility with `releases files <VERSION>` commands.
     pub fn get_release_with_legacy_fallback(&self, matches: &ArgMatches) -> Result<String> {
         if let Some(version) = matches.value_of("version") {
             Ok(version.to_string())
