+++ tests/integration/_cases/help/help-windows.trycmd
@@ -29,6 +29,8 @@ OPTIONS:
 
 SUBCOMMANDS:
     debug-files        Locate, analyze or upload debug information files. [aliases: dif]
+    deploys            Manage deployments for Sentry releases.
+    files              Manage release artifacts.
     help               Print this message or the help of the given subcommand(s)
     info               Print information about the Sentry server.
     issues             Manage issues in Sentry.
